Fix delay time & genv1.0

+++ killservers.py
@@ -1,4 +1,5 @@
 import os
+import os.path
 
 def removeall(l, item):
 	while item in l:
@@ -12,11 +13,19 @@ def microservice(str):
 		'frontend': True,
 	}.get(str, False)
 
+if not os.path.isfile('arq'):
+    print('File does not exist. Execute:')
+    print('sudo lsof -i -P -n | grep LISTEN > arq')
+    exit()
+
 with open('arq') as f:
-	lines = f.readlines()
+    lines = f.readlines()
 
 for l in lines:
-	ret = l.rsplit(' ')
-	removeall(ret, '')
-	if microservice(ret[0]):
-		os.system("kill " + ret[1])
+    ret = l.rsplit(' ')
+    removeall(ret, '')
+    if microservice(ret[0]):
+        print(ret[0])
+        os.system("kill " + ret[1])
+
+os.system('rm arq')
